+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This file is used to list changes made in each version of the openstack-dashboar
 * Bump Chef gem to 11.16
 * Add sensitive flag to local_settings file resource
 * allow override of the temporary directory used for file uploads
+* Set default to use only TLS for SSL. OpenStack security note OSSN-0039
 
 ## 9.1
 * python_packages database client attributes have been moved to the -common cookbook