+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 openstack-dashboard Cookbook CHANGELOG
 ==============================
 This file is used to list changes made in each version of the openstack-dashboard cookbook.
+
+## 9.1.2
+* Set default to use only TLS for SSL. OpenStack security note OSSN-0039
+
 ## 9.1.1
 * Updated Berksfile.lock for the UTF8 issue in common